TPS_Predictor_v2/requirements.txt
# Core ML and Data Science
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Bioinformatics
biopython>=1.81
rdkit-pypi>=2023.3.1
requests>=2.31.0
beautifulsoup4>=4.12.0

# Data Processing
tqdm>=4.65.0
joblib>=1.3.0
h5py>=3.9.0

# Visualization and Analysis
plotly>=5.15.0
umap-learn>=0.5.3

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
